LPS-172314 Add validation to prevent unnecessary DNS requests
Calling InetAddress.getByName with a 'textual' hostname, results in a DNS request.

This logic was introduced by LPS-106335 and based on the commit message, it was intended to handle portal URLs that use an IP Address
- LPS-106335 Check portal URL not only by hostname but by IP address as well (4b21829)

However, we would like to also retain the current behavior, which now also validates URLs that use a hostname associated with the local machine (e.g. localhost + other aliases).
- Providing the additional validation using PortalImpl.isValidPortalDomain would allow users to fine-tune this validation by using the associated portal property