LPS-157425 Use multi-field to re-add support for keyword type for field: *_<languageId>*; needed for wildcard filtering in Headless API
A breaking change was introduced by LPS-152937, which affected *_<languageId>_sortable and *_<languageId>_String_sortable fields.

For the EntityModels in Headless API, there has been a common pattern to use these fields for both filtering and sorting when using the StringEntityField.
- After the introduced change, these fields can no longer be used for filtering, since the type changed from keyword to icu_collation_keyword
- Also, the *_sortable fields in general, were not intended to be used for filtering

This change will reintroduce the fields with keyword type, in the format: *_<languageId>.keyword