LPS-189944 Patch DisconnectCommand to avoid InterruptedException from being shown in the console for a telnet connction, when running the disconnect command
Fix: Before executing the closeable for a TelnetConnection, change the status of the associated job
from foreground to background, to allow the job to finish gracefully.

Issue: The job that is executing the DisconnectCommand is being monitored: JobImpl.this.wait() and is getting interrupted
- https://github.com/apache/felix-dev/blob/org.apache.felix.gogo.runtime-1.1.2/src/main/java/org/apache/felix/gogo/runtime/CommandSessionImpl.java#L802

Reason: The closeable for TelnetConnection calls: this.interrupt()
- https://github.com/eclipse-equinox/equinox/blob/I20220616-0820/bundles/org.eclipse.equinox.console/src/org/eclipse/equinox/console/telnet/TelnetConnection.java#L115-L123

Fix Result: When the job status is changed from foreground to background, the job associated with the telnet connection
(that is executing the disconnect command), will be notified and will allow the thread to proceed.
This will avoid the InterruptedException from being shown in the console of the telnet connection.