LPS-201574 Rely on servlet request to provide the proper host name, to better support HTTP/2 (h2)
Based on the Servlet Request API documentation, the servlet request
first tries to use the Host header field, which would be similar to
our current logic.

With the newer HTTP/2 protocol, the Host header field is no longer
used and has instead been replaced by the pseudo-header:
":authority".

On Tomcat, httpServletRequest.getServerName() accounts for both
the Host header field and the pseudo-header: ":authority" field,
and returns the proper Host name when using HTTP/1.1 or HTTP/2.