LPD-2110 Exclude groupFriendlyURL from the requestURI if the original requestURI does not include it
This method is only being used by FriendlyURLServlet to generate a
redirect link. Previously, the generated link would always include
the groupFriendlyURL.

But, based on the use-case for LPS-131068, the groupFriendlyURL
should only be included if the original request includes it.

The approach for this fix:
- Attempt to retrieve the groupFriendlyURL for the
  OriginalHttpServletRequest
- If the OriginalHttpServletRequest does not have a
  groupFriendlyURL, exclude it from the current requestURI that
  would eventually be used for the generated link

This also fixes the issue for LPD-2110, since it's a regression
bug that was introduced by the original fix for LPS-131068
(which we are reverting in this PR).