LPD-20262 import org.apache.xerces.util to prevent the WARN message for org.xml.sax.SAXNotRecognizedException: Property 'http://www.oracle.com/xml/jaxp/properties/entityExpansionLimit' is not recognized.
The SAXNotRecognizedException is being thrown during interactions
between the third-party libraries: xmlbeans and XercesImpl.

xmlbeans supports implementations from both: XercesImpl and JDK.

When supporting the implementation for XercesImpl, there is logic
to retrieve the class: org.apache.xerces.util.SecurityManager.

However, when we modularized and created shared-dependencies-poi,
we did not provide access to this class, resulting in a swallowed:
java.lang.ClassNotFoundException:
org.apache.xerces.util.SecurityManager

Due to the ClassNotFoundException, xmlbeans eventually fails to
configure the EntityExpansionLimit and logs the mentioned WARN
message.