LPD-21945 Add ElResolver to utilize flag to optimize performance for El Expressions
This usage is based on a performance optimization introduced by
Tomcat:
- Bug ticket: https://bz.apache.org/bugzilla/show_bug.cgi?id=57583
- Commit: https://github.com/apache/tomcat/commit/887780699806b5e72f601fc222e4b9c720e61d92

Note: We are utilizing the flag that was introduced in the previous
commit by adding a custom ElResolver. This avoids having to patch
another class in another third-party library
(javax.servlet.jsp:jsp-api).

The approach taken for this ElResolver is to shortcut the
ScopedAttributeELResolver, based on the referenced optimization
and the existing logic of the ScopedAttributeELResolver being used.

Note: The ScopedAttributeELResolver that is being used for our
current environment is based on the third-party library:
javax.servlet.jsp:jsp-api
- Reference: https://github.com/javaee/javaee-jsp-api/blob/javax.servlet.jsp-api-2.3.2-b01/src/main/java/javax/servlet/jsp/el/ScopedAttributeELResolver.java