POC: Fix possible regression introduced by LPS-13699 and LPS-25275?
Before these two fixes:
- LPS-13699 Asset Publisher Templates (f4a1e54)
- LPS-25275 (f696d94)

The value of companyGroupId would only be greater than 0, when set
with a custom value via a token.

However, after these two fixes, the logic was changed where the
companyGroupId is set based on the themeDisplay and the token now
has a much lower priority.

It appears these two fixes were introduced to add support for
DDM Templates. However, we can see that it affected the logic that
was originally only used by Journal.

It's unclear whether the logic changes were intended and actually
needed for handling the DDM Template use-case. And so, it's
possible this is simply just a regression bug.