POC: Match behavior with JournalTransformer? ; prep next
If we assume the changes from the two tickets actually introduced a
regression bug, then we should also consider whether the logic
for the Transformer class that is handling the DDMTemplates should
also match on how to handle the companyGroupId.

However, I'm not familiar enough with the DDMTemplate component to
decide what should be the proper changes.