LPD-31881 Retrieve exception using our custom attribute key
Changes from LPS-132728 introduced a new attribute key to store
the exception for this error page, to avoid relying on the JSP API
(0c75f7ed).

To further avoid utilizing the JSP API, we will not be using the
variable name: 'exception', since this variable is implicitly used
by the JSP API for error pages. We will instead use a different
variable name.