LPD-31881 Retrieve exception using our custom attribute key
Changes from LPS-132728 introduced a new attribute key to pass the
expected exception for this error page, to avoid having to rely on
the JSP API (0c75f7ed).

This error page will continue to declare itself as a JSP error page
and continue to respsect the variable name: 'exception' that is
implicitly provided for JSP error pages.

In summary, if for some reason an exception is being provided via
the JSP mechanism, we will respect this exception and prioritize it
to be displayed on this error page. Otherwise, we will provide the
exception that is expected to be provided via our custom attribute
key.