LPD-25508 Delay company creation via OSGI config, until initial request
This company creation mechanism was intended to be used with
DBPartition enabled, which expects that the default company is
fully ready before a new partition is created.

And so, delaying the company creation until the initial request
should help to ensure that the default company is fully ready,
before we attempt to create a new company with DBPartition enabled.