LPD-48010 Fix expected locale for test case
The previously declared expected locale was set based on the the
hardcoded layout's locale. However, that was incorrect, since this
test case actually has no connection with the hardcoded layout.

Since the layout is no longer hardcoded, PortalImpl.getLocale
eventually retrieves the locale based on the company's layout,
which results in en_US and would be the expected behavior.