LPD-48010 If a layout has an associated virtual host, consider using it's default language as the locale when locale.prepend.friendly.url.style=0
After LPS-175783 (a45bad474b), VirtualHost decoupled from using the
key: I18N_LANGUAGE_ID to declare it's default languageId by
introducing it's own key: VIRTUAL_HOST_LANGUAGE_ID

However, the usage of the new VIRTUAL_HOST_LANGUAGE_ID key only
took into consideration the logic in I18nFilter. It did not account
for the logic in PortalImpl.getLocale, which also took into
consideration the value of I18N_LANGUAGE_ID, when selecting a
locale.

As a result, a use-case was missed. Specifically when the portal
property: locale.prepend.friendly.url.style=0 is set and a layout
with an associated virtual host has a configured default language.

The goal of this change is to cover the missing use-case. By
setting the locale in the session attribute: LOCALE,
PortalImpl.getLocale will eventually be called and will take into
consideration the set value when selecting a locale.