LPD-51826 Add additional test cases and modify assertion to account for the limitation of the Glowroot API
The Glowroot API cannot precisely return the results based on our
provided startTime, as the tests are executed very quickly.

As a result, transactions from the other test cases may show up
when querying the API, even though we specify a later startTime.
Also, there is no option to sort based on latest transaction, so
we cannot just check the latest entry.

As a workaround, we are checking whether the expected message is
found as one of the entries in the returned transactions.