LPD-53259 Add remaining pluginPackageProperties using analyzer.setProperty instead of analyzer.setProperties
Calling analyzer.setProperties does not make sense for our use-case
since it also introduces additional BND processing logic that
results in unwanted side-effects, such as clearing custom plugin
changes that have been made on the analyzer.

The logic to clear custom plugin changes was added to the library
in a later version. And so, the issue was only observed after
Liferay upgraded to a newer version of the library: biz.aQute.bnd