LPD-56017 Consolidate registering custom TypeConverters in Expando for consistency and to make sure they are made available to both Expando services and JSON services
Previously, some of the custom TypeConverters were prepared within
the static block of JSONWebServiceActionImpl and this class would
be loaded only when the related JSON OSGI component is activated.

This resulted in inconsistent behavior within Expando.

Consolidating the custom TypeConverters to be handled within
Expando would provide a consistent behavior, since the
ExpandoValueLocalServiceImpl is always instantiated during portal
startup.