LPD-60239 Set AbsoluteRedirectsFilter as the first filter
Based on the javadoc, this filter is expected to be the first
filter and to assign the company ID in the request. Making
a call to PortalInstances.getCompanyId the first time, results in
the company ID being set in CompanyThreadLocal.

Scenario for the bug being fixed: the request is handled by the
IgnoreFilter early on the request and is then quickly forwarded to
the utility error page, which now has complex logic, such as
permission checks (LPS-189420/LPD-26807). At this point, the
CompanyThreadLocal has not been set with a company ID.

Permission checks rely on the correct companyId to be set in
CompanyThreadLocal.

With the AbsoluteRedirectsFilter now being the first filter, the
CompanyThreadLocal will now be set earlier on the request.