LPD-58634 Handle checking group permissions separately to isolate the different layers of permission checking; prep next
The _skipReplace method provides the first layer of permission
checking: general permission checks for the resource.

The new _filterResourceViewableGroupIds method provides a second
layer of permission checking: group permission checks for the
resource.

The _getPermissionWherePredicate and _insertResourcePermissionSQL
methods and remaining logic in filter method, provide a third layer
of permission checking: individal permission checks for the
resource.